Refactor: Split BlogDashboard component
Split the BlogDashboard component into smaller, more manageable files.
